Use `snapista` as a backend to use `SNAP`
While leaving some customisation on the table, using `snapista` instead
of `snappy` gives us access to multiprocessing. Indeed, the former
create a `XML` graph and process it using `gpt`, the command line
interface of `SNAP`. And as a consequence, moving out of reach of the
`GIL`, which prevented us from getting decent multiprocessing
performances.

This patch unfortunately drops some feature of `main.py` but they were
mainly useful for development purpose. We don't need them to execute
this on a server.

It also simplifies the installation process as we are now installing
`SNAP` as a conda package. It also gets us rid of the complex
configuration of `snappy`.