Implement models for the database
We are now using this implementation to read the database in
`get_detections`.

The database is noz composed of two tables: 'tiles' and 'detections'.
The first one stores each tile (corresponding to a Sentinel product) and
link to the second table. The latter is responsable for every detection,
and each one point to a tile.

Co-authored-by: Lucas CHOLLET <lucas.chollet@free.fr>